+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Generator/Constants.hs
@@ -1,4 +1,5 @@
-module Test.Cardano.Ledger.Shelley.Generator.Constants (
+module Test.Cardano.Ledger.Shelley.Generator.Constants
+  {-# DEPRECATED "Use `Test.Cardano.Ledger.Shelley.Constants` instead" #-} (
   Constants (..),
   defaultConstants,
 ) where