+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add `certsTxBodyL` to `EraTxBody`
 * Introduce `TxCert` type family and `EraTxCert` type class.
+* Deprecate `Delegation`
 
 ## 1.2.0.0